auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Auth Service"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "hiredis>=3.0.0",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.2.0",
]
